+++ src/index.ts
@@ -7,5 +7,6 @@ declare global {
         clearNewsTickerText: () => void,
         setSliderValue: (e: Element, value: number) => number,
         gainLumps: (lumpsToGain: number) => void,
+        warpTimeToFrame: (frame: number) => void,
     };
 }